+++ e2e/src/common/mod.rs
@@ -1,1 +1,1 @@
-pub mod http;
+pub mod platform_client;
