+++ e2e/src/common/platform_client.rs
